backend/requirements.txt
# Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
django-filter>=23.3
drf-yasg>=1.21.5

# Database and Storage
psycopg2-binary>=2.9.6
Pillow>=9.5.0
django-storages==1.14.2
boto3==1.28.64

# Caching and Task Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# Utils and Helpers
python-dotenv==1.0.0
python-decouple>=3.8
gunicorn==21.2.0
python-magic==0.4.27
django-cleanup==8.0.0
django-environ==0.11.2
django-tagging==0.5.0
openai==1.3.0
requests>=2.31.0
python-jose>=3.3.0
cryptography>=40.0.2
twilio>=8.2.2  # Optional - for phone verification

# Testing Dependencies
coverage==7.6.8
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
